refactored and fixed some method call bugs
